refactor: Modify the folder structure and code according to PR change requests
