+++ src/ui-core/components/atoms/typography/H2/H2.jsx
@@ -1,9 +1,10 @@
 import React from "react";
 import "../Typography.css";
 
-export default function H1({ children }) {
+export default function H1({ children,className }) {
+  console.log(className);
   return (
-    <h2 className="main-heading">
+    <h2 className={className}>
       {children}
     </h2>
   );
