+++ src/ui-core/components/atoms/typography/Typography.css
@@ -4,7 +4,4 @@
   color: #0e0e0e;
   text-decoration: none !important;
 }
-
-.sub-heading{
-    
-}
+
