+++ src/ui-core/components/molecules/ItemCard/ItemCard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-
+import H2 from "../../atoms/typography/H2/H2";
 import "./ItemCard.css";
 
 export default function ItemCard(props) {
@@ -7,14 +7,14 @@ export default function ItemCard(props) {
 
   return (
     <div className="item-card-container">
-      <h2 className="item-card-title">{title}</h2>
+      <H2 className="item-card-title">{title}</H2>
       <img src={image} alt="Flash sale item" className="item-card-img" />
       <div
         className={`item-card-details-container ${
           category === "men's clothing" ? "card-mens" : "card-womens"
         }`}
       >
-        <h2 className="item-card-price">Rs {price}</h2>
+        <H2 className="item-card-price">Rs {price}</H2>
         <p className="item-card-desc">{description}</p>
       </div>
     </div>
