+++ src/ui-core/templates/Category/Category.css
@@ -1,10 +1,3 @@
-.category-container-titile{
-
-}
-.category-main-container{
-     
-}
-
 .category-card-container:first-child{
     background-color:#2BD9AF ;
 }
