+++ src/ui-core/templates/Category/Category.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import "./Category.css";
-import FlexContainer from "../../layouts/FlexContainer/FlexContainer";
+import SectionLayout from "../../layouts/SectionLayout/SectionLayout";
 import CategoryCard from "../../components/molecules/CategoryCard/CategoryCard";
 
 export default function Category() {
   return (
     <div className="category-main-container"> 
-      <FlexContainer heading="Categories">
+      <SectionLayout heading="Categories">
         <CategoryCard name="Men's Clothing" page="mens" />
         <CategoryCard name="Women's Clothing" page="womens" />
-      </FlexContainer>
+      </SectionLayout>
     </div>
   );
 }
