--- src/ui-core/templates/FlashSale/FlashSale.css
@@ -1,1 +1,0 @@
-
