--- src/ui-core/templates/pages/HomePage/HomePage.css
@@ -1,3 +1,0 @@
-.home-page-containerv{
-    
-}
