--- src/ui-core/templates/pages/MensClothing/MensClothing.css
@@ -1,10 +1,0 @@
-.mens-clothing-sub-container{
-    margin:40px;
-}
-.mens-clothing-item-container{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-/* to layout */
