--- src/ui-core/templates/pages/NotFound/NotFound.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function NotFound() {
-  return (
-    <div>
-      Page not found!
-    </div>
-  )
-}
