--- src/ui-core/templates/pages/WomensClothing/WomensClothing.css
@@ -1,8 +1,0 @@
-.womens-clothing-sub-container{
-    margin:40px;
-}
-.womens-clothing-item-container{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
